[TS-25] Feat: chattingMessage
