+++ src/constants/chatData.tsx
@@ -18,7 +18,7 @@ export const chatData = [
 			`어떤 습관을 함께 만들어 볼까요?`,
 			`매일 해도 무리 없는 쉬운 것부터 시작하기를 추천해요 😊`,
 		],
-		reply: ["습관 선택"],
+		replyBtnMessage: ["습관 선택"],
 	},
 	{
 		id: "habit",
@@ -27,7 +27,7 @@ export const chatData = [
 			`이번엔 정체성을 정해 볼게요`,
 			`${userData.habit}를(을) 통해서 ${userData.nickName}님은 어떤 사람이 되고 싶으세요?`,
 		],
-		reply: ["정체성 선택"],
+		replyBtnMessage: ["정체성 선택"],
 	},
 	{
 		id: "identity",
@@ -35,7 +35,7 @@ export const chatData = [
 			`좋습니다 :)`,
 			`지금 이 순간부터 ${userData.nickName}님은 ${userData.identity} 사람이에요`,
 		],
-		reply: ["그런데 정체성을 왜 정하는 거야?"],
+		replyBtnMessage: ["그런데 정체성을 왜 정하는 거야?"],
 	},
 	{
 		id: "reasonIdentity",
@@ -44,7 +44,7 @@ export const chatData = [
 			`그리고 우리는 놀랍도록 스스로가 믿는 대로 행동하죠.`,
 			`오늘부터 진심으로 ${userData.nickName}님은 ${userData.identity} 사람이라고 믿어 주세요!`,
 		],
-		reply: ["알겠어!"],
+		replyBtnMessage: ["알겠어!"],
 	},
 	{
 		id: "time",
@@ -61,7 +61,7 @@ export const chatData = [
 			`먼저 시간을 정해 볼게요!`,
 			`다른 일에 방해 받지 않는 시간 혹은 매일 지키기에 수월한 시간으로 설정해 주세요`,
 		],
-		reply: ["시간 선택"],
+		replyBtnMessage: ["시간 선택"],
 	},
 	{
 		id: "location",
@@ -70,7 +70,7 @@ export const chatData = [
 			`어떤 장소가 ${userData.nickName}님에게 매력적인가요?`,
 			`가는 것만으로 만족스럽거나 습관을 쉽게 할 수 있는 곳으로 명확하게 적어 주세요`,
 		],
-		reply: ["장소 선택"],
+		replyBtnMessage: ["장소 선택"],
 	},
 
 	{
@@ -81,17 +81,17 @@ export const chatData = [
 			`예를들어 책을 읽는다면 30분보다 3장 처럼요 :D`,
 			`3장으로 시작했지만, 점차 자리 잡으면 30장도 쉬운 일이 되어 있을 거예요!`,
 		],
-		reply: ["행동 선택"],
+		replyBtnMessage: ["행동 선택"],
 	},
 	{
 		id: "alert",
 		message: [`약속을 기억하고 실천을 기록하실 수 있도록 하루 두 번, 알림을 보내드릴게요!`],
-		reply: ["1차 알림 변경", "2차 알림 변경"],
+		replyBtnMessage: ["1차 알림 변경", "2차 알림 변경"],
 	},
 	{
 		id: "organize",
 		message: [`지금까지 나눈 이야기를 정리해 보여드릴게요`],
-		reply: [
+		replyBtnMessage: [
 			"정체성 변경",
 			"시간 변경",
 			"장소 변경",
@@ -107,7 +107,7 @@ export const chatData = [
 			`습관 실천이 가장 핵심이지만, 기록 역시 정말 중요해요 :)`,
 			`무엇 때문에 쉬었는지 혹은 오늘 실천은 어땠는지 매일 제게 알려주세요!`,
 		],
-		reply: ["알겠어"],
+		replyBtnMessage: ["알겠어"],
 	},
 	{
 		id: "last",
@@ -121,6 +121,6 @@ export const chatData = [
 			`고생 많으셨습니다!`,
 			`${userData.nickName}를 꾸준히 하는 ${userData.nickName}님을 응원할게요XD`,
 		],
-		reply: ["나도 잘 부탁해"],
+		replyBtnMessage: ["나도 잘 부탁해"],
 	},
 ];
