--- src/pages/ChatPage.style.ts
@@ -1,6 +1,0 @@
-import { css } from "@emotion/react";
-
-export const chatStyle = css`
-	max-width: 360px;
-	margin: 0 auto;
-`;
